Added test view
